+++ mesh/3d.h
@@ -14,8 +14,8 @@ namespace mesh{
 
 // GENERICS
   struct Vertex3D {
-    float x, y, z;
-    Vertex3D(float x, float y, float z);
+    double x, y, z;
+    Vertex3D(double x, double y, double z);
     Vertex3D();
     double magnitude();
     bool operator==(const Vertex3D& other) const;
